+++ Common/Label.cpp
@@ -5,14 +5,14 @@
 Label::Label() : Control(){ this->isFocus = false; }
 
 Label::Label(string value) : Control(), value(value) {//the text that it get and set,defualt size of the label by text we get
-	this->setWidth(value.size());
+	this->setWidth((short)(value.size()));
 	this->setHeight(1);
 	this->isFocus = false;
 };
 
 Label::Label(string value, Color background, Color foreground) : Control(), value(value) {//adding a background and frame color and set by the size
 	setColor(background, foreground);
-	this->setWidth(value.size());
+	this->setWidth((short)(value.size()));
 	this->setHeight(1);
 	this->isFocus = false;
 
@@ -20,7 +20,7 @@ Label::Label(string value, Color background, Color foreground) : Control(), valu
 
 Label::Label(string value, Color background, Color foreground, short left, short top) : Control(left, top), value(value) {//inheratence from the interface control to now where to put the label
 	setColor(background, foreground);
-	this->setWidth(value.size());
+	this->setWidth((short)(value.size()));
 	this->setHeight(1);
 	this->isFocus = false;
 
@@ -33,9 +33,8 @@ string Label::getValue(){
 }
 
 void Label::setValue(string value){
-	this->setWidth(value.size());
+	this->setWidth((short)(value.size()));
 	this->setHeight(1);
-
 	this->value = value;
 }
